Change type of token `Num` to `u32`
Based on the specification sheet of MIXAL and MIX, which can be found here:
https://www.gnu.org/software/mdk/manual/mdk.pdf
numbers can only be 31 bits (includes the sign bit).

This is because in MIX, a byte consists of 5 bits and a word consists of 5 bytes and a sign bit, so 31 bits in total.

The closest type that is supported by Rust is `u32`, so we update the type of `Num` token to be as consistent as possible with MIX.

+++ src/lexer.rs
@@ -89,7 +89,7 @@ pub enum Token {
     #[regex("[a-zA-Z]([a-zA-Z]|[0-9]|_)*", to_string)]
     Id(String),
     #[regex("([1-9][0-9]*)|0", to_num)]
-    Num(i64),
+    Num(u32),
 
     // The following variant is not a token.
     // We use it as a value in AST nodes
@@ -103,6 +103,6 @@ fn to_string(lex: &mut Lexer<Token>) -> Option<String> {
     Some(string)
 }
 
-fn to_num(lex: &mut Lexer<Token>) -> Option<i64> {
+fn to_num(lex: &mut Lexer<Token>) -> Option<u32> {
     Some(lex.slice().parse().ok()?)
 }
